refactor: post detail dialog 를 widget으로 분리
