+++ src/widgets/post/detail-dialog/index.tsx
@@ -1,10 +1,10 @@
 import { Edit2, Plus, ThumbsUp, Trash2 } from "lucide-react"
-import { useDialogStore } from "../../store/dialog"
-import { useSelectedPost } from "../../store/selectedPost"
-import { useSearchQueryStore } from "../../store/searchQuery"
-import { highlightText } from "../../shared/lib/highlightText"
-import { Button, Dialog, DialogContent, DialogHeader, DialogTitle } from "../../shared/ui"
-import { useComments } from "../../entity/comment/model/useComments"
+import { useDialogStore } from "../../../store/dialog"
+import { useSelectedPost } from "../../../store/selectedPost"
+import { useSearchQueryStore } from "../../../store/searchQuery"
+import { highlightText } from "../../../shared/lib/highlightText"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
+import { useComments } from "../../../entity/comment/model/useComments"
 
 export const POST_DETAIL_DIALOG = "POST_DETAIL_DIALOG"
